requirements.txt
fastapi==0.128.0
uvicorn==0.22.0
python-multipart==0.0.6
requests==2.32.0
mysql-connector-python==9.5.0
bcrypt==4.2.1      
python-jose==3.3.0  
psycopg2-binary==2.9.11    